pyproject.toml
[tool.poetry]
name = "platform-for-paid-content"
version = "0.1.0"
description = ""
authors = ["KateRomanova <kate.romanova@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"
black = "^24.8.0"
pillow = "^10.4.0"
django-phonenumber-field = "^8.0.0"
django = "4.2.2"
djangorestframework-simplejwt = "^5.3.1"
stripe = "^10.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
